+++ tests/Core/Connector/AuthRequestFinalizerTest.php
@@ -3,7 +3,7 @@
 namespace Smolblog\Core\Connector;
 
 use PHPUnit\Framework\TestCase;
-use Smolblog\Core\Command\CommandBus;
+use Smolblog\App\CommandBus;
 
 final class AuthRequestFinalizerTest extends TestCase {
